backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
supabase==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Solana & Blockchain
solana==0.34.0
solders==0.20.0
anchorpy==0.19.1
base58==2.1.1

# HTTP & API
httpx==0.24.1
aiohttp==3.9.1
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# WebSocket
websockets==12.0

# Social Media APIs (optional for now)
# tweepy==4.14.0
# discord.py==2.3.2

# Data Processing & Analysis
pandas==2.1.4
numpy==1.25.2
ta==0.10.2
ccxt==4.1.64

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3 